refactor: 관리자 페이지 및 상품, 쿠폰 관리 컴포넌트 분리
